+++ src/migrations/Install.php
@@ -1,41 +1,43 @@
 <?php
-namespace selvinortiz\shield\migrations;
+namespace verbb\shield\migrations;
 
 use craft\db\Migration;
 
-/**
- * Class Install
- *
- * @package selvinortiz\shield\migrations
- */
 class Install extends Migration
 {
+    // Public Methods
+    // =========================================================================
+
     /**
      * @inheritdoc
      */
-    public function safeUp()
+    public function safeUp(): bool
     {
         $this->createTable('{{%shield_logs}}', [
-            'id'          => $this->primaryKey(),
-            'type'        => $this->string(25),
-            'email'       => $this->string(),
-            'author'      => $this->string(100),
-            'content'     => $this->text(),
-            'flagged'     => $this->boolean(),
-            'ham'         => $this->boolean(),
-            'spam'        => $this->boolean(),
-            'data'        => $this->text(),
+            'id' => $this->primaryKey(),
+            'type' => $this->string(25),
+            'email' => $this->string(),
+            'author' => $this->string(100),
+            'content' => $this->text(),
+            'flagged' => $this->boolean(),
+            'ham' => $this->boolean(),
+            'spam' => $this->boolean(),
+            'data' => $this->text(),
             'dateCreated' => $this->dateTime()->notNull(),
             'dateUpdated' => $this->dateTime()->notNull(),
-            'uid'         => $this->uid(),
+            'uid' => $this->uid(),
         ]);
+
+        return true;
     }
 
     /**
      * @inheritdoc
      */
-    public function safeDown()
+    public function safeDown(): bool
     {
         $this->dropTableIfExists('{{%shield_logs}}');
+
+        return false;
     }
 }