+++ src/models/Settings.php
@@ -1,25 +1,32 @@
 <?php
-namespace selvinortiz\shield\models;
+namespace verbb\shield\models;
 
 use craft\base\Model;
 
 class Settings extends Model
 {
-    public $akismetApiKey             = '';
-    public $akismetOriginUrl          = '';
-    public $logSubmissions            = false;
-    public $enableContactFormSupport  = true;
+    // Properties
+    // =========================================================================
+
+    public $akismetApiKey = '';
+    public $akismetOriginUrl = '';
+    public $logSubmissions = false;
+    public $enableContactFormSupport = true;
     public $enableGuestEntriesSupport = true;
-    public $enableSproutFormsSupport  = false;
-    public $enableCommentsSupport     = false;
+    public $enableSproutFormsSupport = false;
+    public $enableCommentsSupport = false;
 
-    public function rules()
+
+    // Public Methods
+    // =========================================================================
+
+    public function defineRules(): array
     {
-        $rules = [
-            [['akismetApiKey', 'akismetOriginUrl'], 'required'],
-            [['akismetOriginUrl'], 'url'],
-        ];
+        $rules = parent::defineRules();
 
-        return array_merge(parent::rules(), $rules);
+        $rules[] = [['akismetApiKey', 'akismetOriginUrl'], 'required'];
+        $rules[] = [['akismetOriginUrl'], 'url'];
+
+        return $rules;
     }
 }