+++ src/records/Log.php
@@ -1,0 +1,15 @@
+<?php
+namespace verbb\shield\records;
+
+use craft\db\ActiveRecord;
+
+class Log extends ActiveRecord
+{
+    // Public Methods
+    // =========================================================================
+
+    public static function tableName(): string
+    {
+        return '{{%shield_logs}}';
+    }
+}